add useContext for all employees
