+++ frontend/src/components/Buttons/AddButton.jsx
@@ -1,21 +1,14 @@
-import * as React from "react";
+import React, { useContext } from "react";
 import { Button } from "@chakra-ui/react";
 import AddModal from "../Modal/AddModal";
+import { employeesContext } from "../../Providers/EmployeesProvider";
 
-const AddButton = ({
-    employees,
-    setEmployees,
-    isOpen,
-    onOpen,
-    onClose,
-    modalState,
-    setModalState,
-}) => {
+const AddButton = ({ isOpen, onOpen, onClose, modalState, setModalState }) => {
+    const { employees, setEmployees } = useContext(employeesContext);
+
     function onClickHandler() {
         onOpen();
         setModalState("Add");
-
-        console.log("MODAL STATE: ", modalState);
     }
 
     return (
