+++ frontend/src/index.js
@@ -3,10 +3,13 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import { ChakraProvider } from "@chakra-ui/react";
+import EmployeesProvider from "./Providers/EmployeesProvider";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
     <ChakraProvider>
-        <App />
+        <EmployeesProvider>
+            <App />
+        </EmployeesProvider>
     </ChakraProvider>
 );
